Integrate nano with Compass.
The content of the Rules file is ok (routes might be
changed in the future, of course).

The compass/config.rb file contains an add_import_path
to load SCSS files from Foundation. The project_path
is mandatory because we have put config.rb in a
subdirectory (I will move it later).