Add Font Awesome 4.0.3.
To install (manually), follow the instructions at

    http://fortawesome.github.io/Font-Awesome/get-started/

§PRO: Custom LESS or SASS.
(There is no need to change @fa-font-path.)

Put Font Awesome into the static folder, so that
nanoc won't complain about files with different
extension but identical name (we could put Font
Awesome in the root folder, but if we start putting
components there, we are going to pollute the root folder).

To update Font Awesome, simply replace the content
of static/font-awesome (only the fonts and scss
folders are needed).