Modify front matter of sample posts.
All attributes are optional (if the article does
not have a `title` attribute, however, no title
will be printed in the output). If `kind` is omitted,
it is set to `article` during the pre-processing phase.
If `created_at´ is omitted, then it is set to the
current date on preprocessing. If `excerpt` is
omitted, but a `<!-- more -->` separator is present
in the post (see Simple Post), then that will be
used to delimit the summary. The default author and
contact URL defined in nanoc.yaml can be overridden,
too, on a per-post base.