Enable syntax colouring with coderay gem.
The :colorize_syntax filter requires Nokogiri.
But I refuse to install >120MB worth of gems to
get syntax colouring (or to build a static site,
for that matters). The ~30MB currently occupied by the
gems are already an indecent.

(Please put those people to program with a computer of the '80s!)

Fortunately, syntax colouring works out-of-the-box
with kramdown if the coderay gem is installed. And
it does not require the workaround described here:

    https://github.com/nelsonjchen/nanoc-foundation-blog/blob/master/Rules

to modify the language tag so that the filter can
recognize it:

    class MarkdownCodeFenceColorize < Nanoc::Filter
      identifier :markdown_code_fence_colorize_adapter

      def run(content, params={})
        content.gsub(/```(?<lang>.+)/, '```language-\k<lang>')
      end
    end

See also: http://mindspill.net/computing/web-development-notes/nanoc-notes/