Use Pygments for syntax highlighting.
Add the pygmentize gem.
Use Pygments instead of Coderay in blog rules.
Add a sample CSS to style Pygments’s output. The
CSS was created with:

    pygmentize -f html -S colorful -a code
      >content/assets/stylesheets/colorful.css

Add a partial to include the CSS in the pages
that need it. Override a couple of Foundation’s
CSS rules for a better output. Those will be
probably moved in a SCSS override file later.